% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MODIS_alerts.R
\name{filter_MODIS_alerts}
\alias{filter_MODIS_alerts}
\title{Filter MODIS fire alerts}
\usage{
filter_MODIS_alerts(spatial_units_shp, column_units_name, MODIS_shp,
  column_date_name, column_frp_name, column_confidence_name, min_frp,
  min_confidence, size_jpg = c(10, 10), output_folder)
}
\arguments{
\item{spatial_units_shp}{Shapefile filename. This is the shapefile of spatial units to analyze with SATA alerts. It should be projected in UTM WGS84 coordinates system.}

\item{column_units_name}{String. The name of the column indicating the names of the spatial units to analize. Case-sensitive applies.}

\item{MODIS_shp}{Shapefile filename. The shapefile of MODIS alerts from SATA (as points).}

\item{column_date_name}{String. The name of the column indicating the date of alerts. Case-sensitive applies.}

\item{column_frp_name}{String. The name of the column indicating the radiative power of alerts (normally as frp). Case-sensitive applies.}

\item{column_confidence_name}{String. The name of the column indicating the confidence of alerts (normally as confidence). Case-sensitive applies.}

\item{min_frp}{Numeric. A value defining the minimun radiative power in filtering.}

\item{min_confidence}{Numeric. A value defining the minimun confidence in filtering.}

\item{size_jpg}{Numeric. Two values which define the width and height of boxplot (in cms).}

\item{output_folder}{Foldername. A folder to use for store alerts filtered.}
}
\value{
Filtered alerts and plots are stored in: *output_folder*.
}
\description{
This function filters MODIS fire alerts. It uses a shapefile of spatial units to generate plots and visualize some statistics of alerts by spatial units.
}
\examples{
#Defining variables from spatial units shapefile
spatial_units_shp <- "C:/DATA/GAF/demo/shp/unidades_espaciales/provincias.shp"
column_units_name <- "dpa_despro" #case sensitive

#Defining variables from MODIS shapefile
MODIS_shp <- "C:/DATA/GAF/demo/shp/alertas/pts_calor_MODIS_30d.shp"
column_date_name <- "acq_date" #case sensitive
column_frp_name <- "frp" #case sensitive
column_confidence_name <- "confidence" #case sensitive
min_frp <- 10 #same units as area column
min_confidence <- 50 #with same units as area column

#Defining variables of ouputs
size_jpg <- c(40,20) #first value defines width and second height
output_folder <- "C:/DATA/GAF/demo/estadisticas"

#running the function
filter_MODIS_alerts(spatial_units_shp = spatial_units_shp,
                            column_units_name = column_units_name,
                            MODIS_shp = MODIS_shp,
                            column_date_name = column_date_name,
                            column_frp_name = column_frp_name,
                            column_confidence_name = column_confidence_name,
                            min_frp = min_frp,
                            min_confidence = min_confidence,
                            size_jpg = size_jpg,
                            output_folder = output_folder)
}
